Make QubitSearchEngine an abstract class
I'm still unclear on the purpose of QubitSearchEngine. I'd like to
define a number of requirements and some basic funcitonality like
logging for search engine implementors (like arElasticSearchPlugin)
but I don't know if the extra work is worth it. Anyways, an abstract
class seems to fit better than an interface in this case.