Set dynamic option to strict, map i18n fields
I thought that it would be a good practice to set the dynamic mode
in 'strict' for both root and inner objects. There is no performance
overhead in objects being dynamic but it is a safe mechanism so
malformed objects won't index data that we don't want to be indexed.
This strategy forced me to map any i18n field available, using
TablleMap classes for that. I also did some cleanups.