Refactor search/autocomplete using ES _msearch
Before, any keystroke in the autocomplete field would query the
elasticsearch server four times since we are looking for four
different types. This commit uses _msearch, which is not yet
implemented in elastica, although they are discussing how and
there is already some code that it was useful while writing this.