Add dateTypes test
I'm trying to understand how the ORM manages the temporal columns
internally. It seems that DateTime objects are accepted during
property assignation and they will be converted to strings using
the param() method (http://git.io/ftlt6w).

In fact, the ORM uses DateTime objects when the createdAt and
updatedAt colums are assigned internally. That can be a bit confusing.
See http://git.io/izf2Qw. It would not be easier to use string?