Stop using __get() for 'notes' in QubitUser
IN SVN r11089, I used __get() for reading the notes of an user
using the property name "notes". This change produced a property
name clase since QubitObject uses the same property in a different
context. As QubitUser extends QubitObject, the former was telling
to the latter that the notes linked by note.user_id should be
stored by note.object_id, where the foreign key was the user id.

Unfortunately, this have caused the loss of notes for some users
if they have been edited. These notes are still in the database
but their original resource id can't be restored unless the
administrator keep backups.

Refs #4539.