Obtain original name, fixes #1593
This commit adds a method called getOriginalName that finds the
corresponding filename for each object so it is assigned as the
title of the archival description.

Also, getMainDmdSec has been updated to be compatible with multiple
structMap elements in the document (physical and logical layouts).

A few logging calls have been added too.