Move app.yml with defaults to config/app.yml
The cascade mechanisms merges the configuration files according
to the following order:
  - /apps/qubit/config/app.yml
  - /config/app.yml

So we are storing some defaults in the latter. The user can decide
which one to edit, ideally the user edits the former as it's the one
ignored by the SCM, in cases where the app is deployed using git.