Improvements in the browser searches
Applying .inline-search along every search found in the browser pages. It uses
mostly the default Bootstrap 2.x style, but customized in ArchivesCanada as its
whites and greys are too distracting in combination with the purple background.

Making some changes in the way that the sort settings are gathered and removed
the table header sorting options in favor of the top-right dropdown.

Fixes #5721.