Refactor cbd/renderer, import dagreD3
Extending Renderer from dagre-d3 wasn't giving me enough flexibility. One
of the main problems was that I want the graph to have a layout strictly based
in the existing hierarchical relationships, but ignoring associative
relationships. dagre.layout won't allow me to do that (and that's something I
want to talk about with Chris, dagre's lead developer). So thew new rendered,
pretty much a copy of the one found in dagre-d3, will pass to dagre.layout a
version of the directed graph limited to hierarchical relationships.

dagre-d3 is not a dependency anymore. dagre is now a dep, so we can still have
access to dagre.graphlib and dagre.layout.

It's work in progress.