Allow partial foreing types in the ES mapping
Gives the posibility to add only a few fields from foreing types
using properties to indicate the normal fields and _i18nFields for the
language fields, the id and sourceCultere are added when building the index.
Include part_of in aip type using this method, with only id and i18n titles.

Conflicts:

	plugins/arElasticSearchPlugin/config/mapping.yml
	plugins/arElasticSearchPlugin/lib/model/arElasticSearchAipPdo.class.php

Conflicts:

	plugins/arElasticSearchPlugin/config/mapping.yml
	plugins/arElasticSearchPlugin/lib/model/arElasticSearchAipPdo.class.php