Mixin to convert dates to ISO8601
It also includes a hack that makes possible to work with non-UTC dates
in ES. We assume the fact that we are pushing localtimes to Elasticsearch.
Elasticsearch returns ISO8601 suffixed with "z" = UTC. We're just removing
the suffix and adding the local timezone date suffix, until we have time
to update our documents and our build scripts to start using UTC dates in
Elasticsearch. FYI: MySQL also uses localtimes.