Override NS and descendant queries fixes #6978
Previously we removed the nested set related code for digital objects
in BaseDigitalObject. We also changed the descendants query to
reflect this and use parent_id for the query instead of lft/rgt.

However, if the user regenerates the database model via propel, the
removed / modified code will revert back due to the generator not
reflecting the changes.

This fix will override the nested set / descendant methods in
QubitDigitalObject, and simply do nothing when they're called.
This should persist / solve the issue if the user ever regenerates
the model.