Support for params in cache_engine
cache_engine used to be a string containing the name of the class used in our
internal cache, e.g. sfAPCCache, sfAPCuCache, sfMemcacheCache, etc...
This commit adds support for a dictionary (class and param keys) in addition
to the old single-string format. This is very useful when we need to pass extra
settings to the cache engine, e.g.

New format:

  cache_engine:
    class: sfMemcacheClass
    param:
      [dict: see sfMemcacheClass for more details]

Old format (still supported):

  cache_engine: sfAPCCache