Hide recover button if successfully recovered.
Added, to the last recovery information included in the fixity status REST
endpoint, whether or not the recovery attempt is current in relation to
the failure report. If the recovery attempt happened after the failure was
reported, it's considered current. If the recovery attempt is current and
successful, we don't show the recover button in the dashboard.

Also fixed issue with global failure status indicator in dashboard reporting
outstanding fails even when AIPs that fail checks are successfully recovered.