Changed MODS bulk export so children export.
Changed the query for information object so if an export's format is
MODS, and the current level only option isn't flagged, child
descriptions are exported in addition to top-level descriptions.

Removed typo in MOD XML template and changed template so publication status
doesn't get shown in title of parents/children.

Updated relatedItem XML to conform to MODS standard.