Fix def. namespaces in EAD breaking import #7280
- Apparently DOMXPath objects require prefixes on *all* namespaces
  when querying with XPath if they are defined, even if it's a
  default namespace. (e.g. xmlns="..." instead of xmlns:xsi="...")

- You need to create a dummy prefix for default namespaces when you
  register the namespace, see: http://stackoverflow.com/a/6475568 at
  the end of the top answer.

- You cannot remove default namespaces using DOMElement or DOMDocument

The solution I came to is instead of passing $doc->load($filename),
we read in the filename with file_get_contents, remove the default
namespace from the <ead...> tag using regex, then call $doc->loadXML(...)
on the resulting string instead. Since load() was reading the file in
anyway, this performance hit shouldn't be bad.

I also set preg_replace's option to only replace *1* occurrence,
so it should stop parsing the string after only the first few lines.

I felt this was the best solution that while a bit dirty, didn't
hurt performance and made AtoM so it doesn't break on all those
sites that were using the broken 2.0.2 EAD export.