Add output, file moving to XML export. refs# 7627
When in verbose mode, show name of file being imported. This makes it
easier to pinpoint when in import data problems exist if an import
stops due to an error.

Also added option to move successfully imported files into a
separate directory, so if an import stops due to an error you don't
have to start from scratch.