Add top-level description filter, refs #7570
The user can now use a radio button to specify whether to
show top level archival descriptions when searching/browsing,
or to display all descriptions.

When the user is browsing archival descriptions, the default
setting is now to display top level descriptions only (as it
was in ICA-AtoM).

Fix top LOD label [X] not working

Use $this->request instead of passing paramaters