Use $sf_data->getRaw(...) when needed
There are places where the way the code is written just don't work with escaped
objects. For example, getAll() in a parameter holder does not return an array
anymore but an object that implements ArrayAccess. We use the + operator to
append that to another array in multiple templates, but that won't work
anymore.

It seems that using the raw version of the parameter holder is secure in the
cases used in this commit, e.g. for links generation the HTML entities are
still escaped somewhere else.