Remove login options in read only mode, refs #7913
Checks if the app is in read only mode looking at the env. variable
ATOM_READ_ONLY and the setting read_only in app.yml,
giving priority to the env. variable.

If the app is in read only mode, the login button will be removed
and the '.../user/login' url will redirect to the homepage