Add AJAX pagination to repo holdings #8129
Add ability to jump to an arbitrary page

The user will now be able to type in a desired page number in an input box,
and after finishing typing the holdings pager will jump to that page. If the
specified page is invalid, the holdings pager will just re-insert the current
page number into the input box.