Remove unnecessary qubitParentSlug col data #8122
Remove the qubitParentSlug column data when exporting CSVs via the
csv:export task in symfony. This is because we're not gaining anything
by having it there (we already fill out the parentId column), and having
the qubitParentSlug column present was causing errors if AtoM was
generating different slugs than the CSV expected.

This will especially be an issue if we add custom slugs in the future,
if we don't make this change.

Moving forward, the qubitParentSlug column should just be used in CSVs
that want to add additional children to existing records with known slugs.