Respect UI-configured max image width generally
When importing descriptions with digital objects either on the command-line or through the UI, the ui-configured maximum image width setting is not respected. This is likewise true when issuing the digitalobject:regen-derivatives command. It seems the wrong (legacy?) setting is checked.  FWIW, changing reference_image_maxwidth in data/fixtures/settings.yml doesn't help. This change checks the UI value for maxwidth and uses it.