QubitObject: save QubitRelation only if transient, refs #7750
relationsRelatedBysubjectId and relationsRelatedByobjectId are magic getters
that return every single QubitRelation linked. This commit uses its subset
`transient` where only new objects with no representation in the database are
returned.

With this commit we still have a mechanism to postpone the creation of foreign
entities using the ORM, e.g. we can't insert a QubitRelation in the database
unless we postpone that until its corresponding QubitActor is inserted.
The main difference is that with this commit the existing relationships won't
get updated automatically and that responsability passes to the developer.
Fortunately, neither relationsRelatedBysubjectId nor relationsRelatedByobjectId
seem to be used other than for inserts. I've found only one place where calling
save() was necessary and the change was included in this commit.