Fix edge case for year 0000 end dates in ES #8796
If we passed an end date of '0000-00-00' to arElasticSearchPluginUtil::normalizeDateWithoutMonthOrYear(),
we would get to the cal_days_in_month() part in the function and pass it $year='0000', $month='12',
which would result in a return of (bool)false.

Converting (bool)false to a string is blank (''), so the returned 'normalized date' would be
'0000-12-', which is the date in the ES error mentioned in the ticket.

Also rename function normalizeDateWithoutMonthOrYear to normalizeDateWithoutMonthOrDay so the name
is accurate.