Prevent duplicates in /add using autocomplete, refs #8768
In autocomplete.js, the .add sibling learns a new paramameter
data-link-existing="true" used as follows:

<input class="add" type="hidden" data-link-existing="true" value="..."/>

When used, a parameter "linkExisting=true" will be appended to the URI taken
from value="", used in autocomplete.js in order to request the addition of new
items in the database.

This commit also updates the controllers to accept the new parameter and
respond redirecting the user to the existing element (first match found)
instead of creating a new one.

Finally, avoid necessary jQuery constructors in autocomplete.js and update
multipleSelectHasMatches() to exclude the caller when matching repetitions
in select.multiple.