Fix public EAD XML export with drafts, refs #9254
We were relying in the last descendant to close the ancestors <c> tags
and, with the public option, that child may not be included in the
descendants to export. Closing the previous ancestors tags when
we are in the next sibling (rgt value bigger than the previous one) and
making sure all <c> tags are closed at the end allows us to have missing
descendants in the exported hierarchy.