Enhance csv import matching behavior, refs #9995
Added a new checksum column for digital objects on both the import and export
csv files.  This will affect CLI imports when the -update parameter is set.
In this case, if a checksum is present in the input file, and it matches
the checksum in the database, the digital object will not be updated.  If
the checksum is missing or does not match, the digital object will be
loaded as normal.

This change will prevent digital objects from being reloaded on an -update
if they are exact copies of one another.