Improve information object search, refs #10082
- Narrow down how many fields in nested, foreign types inside information object
we search over. Now we only search for things like repository name (but not
history), creator name (but not history), etc. This will ensure we don't
accidentally get weird search results just because a keyword was in a creator's
biog history or things like that.

- Add weighting to search fields when searching, so certain fields yield higher
scores when returning from queries. Also added a 'sort by relevance' option so
we can take advantage of this way of searching.

- Change the default search operator from OR to AND. This will result in more
concise results.