Enhance XML-EAD and CSV import matching behaviour, refs #9989
Added logic in the XML EAD import to try to match the information
object being loaded to a record in the database. If a record is
found, the information object will not be loaded and message
containing the record identifier will be logged. The XML EAD
process did not historically try to match to existing records.

Matching logic for XML EAD import will cascade through different
matching logic:

- first try to match the file name to the keymap table
- if no match, try to match against identifier, title, repo name
- if both fail, no match and record will not be loaded

XML EAD import process also modified to create a keymap record
when an information object is successfully matched and loaded.