Enhance XML-EAD and CSV import matching behaviour, refs #9989
Added logic in the XML EAD import to try to match the information
object being loaded to a record in the database. If a record is
found, the information object will not be loaded and message
containing the record identifier will be logged. The XML EAD
process did not historically try to match to existing records.

Matching logic for XML EAD import will cascade through different
matching logic:

- first try to match the file name to the keymap table
- if no match, try to match against identifier, title, repo name
- if both fail, no match and record will not be loaded

XML EAD import process also modified to create a keymap record
when an information object is successfully matched and loaded.

Added secondary matching logic to the CSV import programs.

The CSV import program first tries to match to a record in the
keymap table (existing logic). This has been enhanced so that
if the keymap logic does not match to a valid record, a second
search will be peformed to try to match the csv record to a
record in the database based on 'identifier', 'title' and
'repository' csv fields. If a record is matched or not matched,
the will follow the same pre-existing logic it would
have when matching against keymap records.