CSV import: issue importing digital objects
An issue was introduced in 9995/6 where digital object
URLs could not be imported: doing so would produce the
error "could not resolve host" in the console. This
change resolves this issue.

Assignment of the url from the CSV file was not
working correctly. This same issue would have
affected loading from path as well.