CSV/XML import matching option, refs #10054
CSV Import:

Modified --update option to be able to take parameters to modify the update
logic. Adding further parameters for --update can leverage work done for this
task. Parameters will be specified as: --update="match". Parameters are
verified and an exception will be thrown if an unknown parameter is specified.

Added logic to implement 'match' parameter with --update option. When set,
Information Objects must match to a record in the database or they will be
skipped and a message will be logged in the console.