Specify csv transform script WebUI, refs #10138
The CSV Import page in AtoM now allows a user to choose if they would like
to run a transformation script against the csv file being uploaded.

The transformation script to be run must be configured in config/app.yml by an
administrator prior to the user being able to see the checkbox on the CSV
Import page. If a transformation script is configured, the full path and
name of the transform script will also display opposite the checkbox on the
CSV Import page.

If the transformation script returns an exit code greater than zero, this
indicates a failure in the transform script and an exception will be thrown
and the csv import will not proceed.

The transform script must take 2 parameters:
1) The file to be transformed
2) The output file name that will contain the tranformed csv

The transform script must return a non-zero exit code to indicate an error
condition. An exit code of zero indicates success and will cause the csv
import to proceed. Any error messages must be output by the script to stdout
or stderr.