Update 'Matching Descriptions' url, refs #10295
In AtoM 2.3, the search was changed to use the informationObject
module. Search URLs changed from:

- "search?query=Dalhousie&repos=" to
- "informationobject/browse?topLod=0&query=Dalhousie&repos="

The autocomplete drop down for "All Matching Descriptions" was not
updated to match the changes. If a user selects it, they will be shown
encoded json instead of a valid AtoM search page.

This change corrects the url & module used when selecting "All
Matching Descriptions".