Update autocompleteSuccess.php
In AtoM 2.3, the search was changed "from search?query=Dalhousie&repos=" to "informationobject/browse?topLod=0&query=Dalhousie&repos=". The autocomplete drop down for "all matching descriptions" was not updated to match the changes. If a user selects it, they will be shown encoded json instead of a valid AtoM search page.