Background bulk export from clipboard, Refs #9998
This commit includes a new CSV/XML export form, adds the ability to export
XML records in the background plus some additional user interface improvements
for the export process.

Background job modifications:

- Created a new background job that wraps the existing XML export CLI
utility to provide asynchronous export capability to the WebUI. XML output
is downloadable from the Job Management page as a zip file.
- Modified the existing CSV export job to allow an option to include
descendant records in the CSV export results.

User interface enhancements:

- a new form accessed via the clipboard page for setting options for
CSV and XML Exporting from the clipboard.
- modifications to the Clipboard page template to include an 'Export'
button and removal of the old 'Export CSV' link.

Clipboard menu dropdown enancements:

- a count of records by entity type is now included in the dropdown menu.

QubitClipboard has been changed in terms of how the clipboard items are
serialized - instead of a flat array of slugs going into sfStorage, the
array is now one level deeper so that slugs are grouped by class name.
This was done to ensure that the class name was determined when an item
is added to storage, as opposed to figuring out the class name of all
objects in the clipboard when viewing the menu.