Create XML Export background job, Refs #9998
Created a new background job that wraps the existing XML export CLI
utility to provide asynchronous export capability to the WebUI. XML output
is downloadable from the Job Management page as a zip file.

Modified the existing CSV export job to allow an option to include
descendant records in the CSV export results.

  Changes to be committed:
	modified:   config/gearman.yml
	modified:   lib/job/arInformationObjectCsvExportJob.php
	new file:   lib/job/arXmlExportJob.php