Clipboard menu items by class name, Refs #9998
This commit includes changes to the Clipboard menu dropdown to display
a count of items in the clipboard by class name.

QubitClipboard has been changed in terms of how the clipboard items are
serialized - instead of a flat array of slugs going into sfStorage, the
array is now one level deeper so that slugs are grouped by class name.
This was done to ensure that the class name was determined when an item
is added to storage, as opposed to figuring out the class name of all
objects in the clipboard when viewing the menu.

 Changes to be committed:
	modified:   apps/qubit/modules/menu/actions/clipboardMenuComponent.class.php
	modified:   apps/qubit/modules/menu/templates/_clipboardMenu.php
	modified:   apps/qubit/modules/user/actions/clipboardToggleSlugAction.class.php
	modified:   js/clipboard.js
	modified:   lib/QubitClipboard.class.php