Enable ACL checks on export, Refs #10000
Historically, ACL checks are disabled for CLI tasks. This applied also to
asynchronous background jobs which had an environment of 'cli'.

This change introduces a new environment type for background gearman workers:
'worker'. All classes extending arBaseJob will now have an enviroment type of
'worker' instead of 'cli'.

In addition, the owner of the worker job will be set as the user in background
workers.

Both the new environment type ('worker') and the fact that the user context is
now being set mean that ACL checks will now run for all background worker
tasks.