New custom logger for jobs, refs #10621
A new logger called arJobLogger is created by arBaseJob with context about the
current running job. This new logger has been created so it can be injected to
other parts of the application where a logger is accepted so we can capture
their events.

arJobLogger also dispatches logs to `gearman.worker.log` as there is an
observer in `jobWorkerTask.class.php` passing these messages to sfTask so
ultimately they're shown in the standard stream.