renderDate leading zeroes on day, Refs #10641
This change addresses an issue with date formatting where if the day had a
leading zero it would be formatted incorrectly when displayed on the screen.
e.g. 2016-12-03 would display as 2016-123

Note that this issue would only occur when the month was a two digit month as
in the example above.

This corrects the issue - month and day values with leading zeroes will not
have them truncated off. The issue where a day value less than zero preceeded
by a month value with two digits, will no longer merge them together as in the
above example.

Note that this function will not add leading zeroes if they are not present.
These date values are stored in the database with leading zeroes present so
should always display as represented in the database.

e.g. 2016-6-9 will display as 2016-6-9
     2016-06-09 will display as 2016-06-09
     2016-12-6 will display as 2016-12-6
     2016-12-06 will display as 2016-12-06