Institutional Scoping enhancement, Refs #10733
Institutional Scoping enhancement for AtoM. When activated, the feature
will allow users to be scoped to a specific repository such that browsing
and searching will restict a user to a specific repository. Repository
theming will apply to pages when scoped to give the user visual cues
indicating they are browsing with a repository. A new search block will
be present on the left hand side from which browse and search will be done.

The User attribute 'search-realm' is used to maintain user context when a
user is or is not scoped to a particular institution. The search-realm
will be set to a repo id when scoped, and will not be set when the user
is globally scoped.

Changes include:

- migration 149 to set up new settings and menu items
- settings and menu item config in menus.yml and settings.yml
- Global Settings page change to add new option to "Enable insitutional
scoping"
- new browse menu component which provides the ability to configure custom
menu items separately from the standard browse menu. Also provides different
text labelling than the standard browse component.
- new holdings component which uses the new browse institution component.
This component is the 'Institutional Block' and is present when institutional
scoping is turned on, and the user is currently scoped to an institution (not
browsing globally). Provides search of information objects within the scoped
repository.
- Top main search box will always search the database globally when this
feature is ON

Supporting technical changes:

- in QubitMenu created a new alias called %currentRealm% which will contain
the current repository search context. This is used by menu items on the
Institutional Block as part of the enable_institutional_scoping feature. The
alias will be populated with search-realm if available, else it will try to
use resource->id from sf_route, else try repos param from request.
- added the method removeAttribute() to the myUser class.

Repository theming:

This enhancement includes two new stylesheet components. Both components will
trigger the background theme for the repository they are part of when the
Institutional Scoping feature is turned on and the user is scoped to a repo.

The application of the Repository theme to these pages is intended to give
visual cues to the user that they are currently scoped to an institution.

These two components differ in their logic to detect when the repository
page style should be applied. Repositories always have the stylesheet applied;
the info obj module should only have it applied if the Institutional Scoping
feature is turned on, and the Achival Description plugin modules have
different display logic for the stylesheet depending on whether the Inst
Scoping feature is turned on or off.

- The information object module component will trigger the theme when the
institutional scoping setting is turned on and the search-realm is set.
This is new behaviour for the information object module and is strictly
dependent on the Institutional scoping feature being turned on.
- The sfIsadPlugin component will first look at the institutional scoping
setting and examine the search-realm for a repo. If institutional scoping
setting is off, it will fall back to the original behaviour for the Archival
Descr plugin modules: try to determine the repo from the sf_route and display
the repo stylesheet if found. This component replaces the use of the repository
stylesheet component in the Archival Description plugin modules and
incorporates new logic for the inst scoping feature.

The following plugins will use the new sfIsadPlugin component:
- arDacsPlugin
- sfDcPlugin
- sfModsPlugin
- sfRadPlugin