Institutional Scoping migration conf, Refs #10733
Institutional Scoping enhancement for AtoM. When activated, the feature
will allow users to be scoped to a specific repository such that browsing
and searching will restict a user to a specific repository. Repository
theming will apply to pages when scoped to give the user visual cues
indicating they are browsing with a repository. A new search block will
be present on the left hand side from which browse and search will be done.

This is the first of several commits to pull this into 2.4.x

This commit includes:
- migration 149 to set up new settings and menu items
- settings and menu item config in menus.yml and settings.yml
- Global Settings page change to add new option to "Enable insitutional
scoping"