Institutional Block components, Refs# 10733
Institutional Scoping enhancements.

Commit includes:

- new browse menu component which provides the ability to configure custom
menu items separately from the standard browse menu. Also provides different
text labelling than the standard browse component.

- new holdings component which uses the new browse institution component.
This component is the 'Institutional Block' and is present when institutional
scoping is turned on, and the user is currently scoped to an institution (not
browsing globally). Provides search of information objects within the scoped
repository.

- custom css to support the browse institution component.