Inst scoping stylesheet component, Refs #10733
This commit includes two new stylesheet components. Both components will
trigger the background theme for the repository they are part of when the
Institutional Scoping feature is turned on and the user is scoped to a repo.

The application of the Repository theme to these pages is intended to give
visual cues to the user that they are currently scoped to an institution.

These two components differ in their logic to detect when the repository
page style should be applied. Repositories always have the stylesheet applied;
the info obj module should only have it applied if the Institutional Scoping
feature is turned on, and the Achival Description plugin modules have
different display logic for the stylesheet depending on whether the Inst
Scoping feature is turned on or off.

- The information object module component will trigger the theme when the
institutional scoping setting is turned on and the search-realm is set.
This is new behaviour for the information object module and is strictly
dependent on the Institutional scoping feature being turned on.

- The sfIsadPlugin component will first look at the institutional scoping
setting and examine the search-realm for a repo. If institutional scoping
setting is off, it will fall back to the original behaviour for the Archival
Descr plugin modules: try to determine the repo from the sf_route and display
the repo stylesheet if found. This component replaces the use of the repository
stylesheet component in the Archival Description plugin modules and
incorporates new logic for the inst scoping feature.